+++ parser_library/src/processing/instruction_sets/low_language_processor.h
@@ -87,15 +87,17 @@ private:
 
     using transform_result = std::optional<std::vector<checking::check_op_ptr>>;
     // transform semantic operands to checking operands - machine mnemonics instructions
-    static transform_result transform_mnemonic(
-        const resolved_statement& stmt, context::dependency_solver& dep_solver, diagnostic_collector collector);
+    static transform_result transform_mnemonic(const resolved_statement& stmt,
+        context::dependency_solver& dep_solver,
+        const context::mnemonic_code& mnemonic,
+        const diagnostic_collector& collector);
     // transform semantic operands to checking operands - default machine instructions
     static transform_result transform_default(
-        const resolved_statement& stmt, context::dependency_solver& dep_solver, diagnostic_collector collector);
+        const resolved_statement& stmt, context::dependency_solver& dep_solver, const diagnostic_collector& collector);
 
     static checking::check_op_ptr get_check_op(const semantics::operand* op,
         context::dependency_solver& dep_solver,
-        diagnostic_collector collector,
+        const diagnostic_collector& collector,
         const resolved_statement& stmt,
         size_t op_position,
         const context::mnemonic_code* mnemonic = nullptr);